behaviour: use oauth http client to download redirect
After doing more thorough acceptance on #97, I realized that the HTTP
request was not using the OAuth HTTP client if the access token was
provided (i.e. if an access token is provided, the initial request will
use it, but if the initial response is a redirect, the subsequent
request will be unauthenticated).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>