behaviour: don't pass Authorization header on redirect
When trying to download release assets that redirected to e.g. AWS S3,
my prior commits would cause an error. Since we were using the OAuth
client Transport as an `http.RoundTripper`, we were always adding the
`Authorization` header (overwriting Go's default behaviour of stripping
the Authorization header on redirects).

To avoid this, just use a vanilla HTTP client and pass the access token
manually. If the initial redirect URL is external to github, check the
Host explicitly to determine whether to add the access token. On
subsequent redirects, this is taken care of by Go's http client
stripping the Authorization header.

I did consider creating a custom `http.RoundTripper` to strip the
Authorization header for any request outside of github.com, but that
felt like overkill here.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>