Fix graphql client creation when only v3 url is given
fixes concourse/concourse#6373

This commit fixes the case where a user only provides this resource with
the `github_api_url` and not the `github_v4_api_url`. In this case the
code was appending "graphql" to the end of the provided GitHubAPIURL.
Seems this only affects github enterprise users depending on how they
configured their v3 and v4 URL's. Apparently something like this was
happening for enterprise users when the only set the github_api_url:

GitHubAPIURL = api.mygit.com/api/v3
would become
GitHubV4APIURL = GitHubAPIURL + "graphql" = api.mygit.com/api/v3/graphql

when they really wanted: api.mygit.com/api/graphql

Another important thing to know is that the Github Client for v3 expects
the API URL to end with a "/". The code already ensures this during
creation of a v3 client.

For graphql it is expected that the v4 API URL does not end in a "/"

Signed-off-by: Taylor Silva <tsilva@pivotal.io>