fix redirect test
The http.Client got smarter. The http client only removes the
Authorization header if the base domain is different. When we start
these two local http test servers, they both have the same "base
domain", which is `127.0.0.1`, so the http client never strips the
Authorization header. To work around this I changed the external server
URL to be `localhost`.

So setup looked like this:

http://127.0.0.1:4455 -> redirect to http://127.0.0.1:5566

Now the redirect looks like this:

http://127.0.0.1:4455 -> redirect to http://localhost:5566

Signed-off-by: Taylor Silva <dev@taydev.net>